// src/app/shared/models/user.ts
export class User {
    name: string;
    email: string;
    token: string;
    password: string;
    phonenumber: string;
    high_qual: string;
    high_qual_perc: number;
    pursuing_status: boolean;
    dob: string;
    isHr: boolean;
    gender: string;
    resume: string;
    picture: string;
    street: string;
    city: string;
    state: string;
    pincode: string;
    verified: boolean;
}